+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+- Bugfix on removing a sounding parcel.
+
 ## [1.14.3] - 2021-05-04
 
 ### Changed